+++ gd-edit-portlet/src/main/java/de/seitenbau/govdata/edit/validator/MapKeys.java
@@ -11,12 +11,25 @@ import javax.validation.Payload;
 @Constraint(validatedBy = MapKeysValidator.class)
 @Target({ElementType.METHOD, ElementType.FIELD})
 @Retention(RetentionPolicy.RUNTIME)
-public @interface MapKeys {
+public @interface MapKeys
+{
+  /**
+   * error message
+   */
   String message();
 
+  /**
+   * allowed keys
+   */
   String[] allowedKeys();
 
+  /**
+   * groups
+   */
   Class<?>[] groups() default {};
 
+  /**
+   * payload
+   */
   Class<? extends Payload>[] payload() default {};
 }