+++ gd-search/src/main/java/de/seitenbau/govdata/search/gui/model/SearchMapViewModel.java
@@ -28,6 +28,7 @@ public class SearchMapViewModel
   private String tileUrl;
   private String geocodingUrl;
   private String credits;
+  private String layers;
   
   
 }